refactor: restructure project into standard Go layout (cmd, internal, ui)
