+++ ui/ui.go
@@ -1,0 +1,6 @@
+package ui
+
+import "embed"
+
+//go:embed static/*
+var StaticAssets embed.FS
